continue ...

+++ Chapter_01/03/routes/events.py
@@ -3,6 +3,7 @@ from fastapi import APIRouter, Depends, HTTPException, Request, status
 from database.connection import Database
 from typing import List
 from models.events import Event, EventUpdate
+from auth.authenticate import authenticate
 
 event_router = APIRouter(
     tags=["Events"]
@@ -26,12 +27,19 @@ async def retrieve_event(id: PydanticObjectId) -> Event:
     return event
 
 @event_router.post("/new")
-async def create_event(body: Event) -> dict:
+async def create_event(body: Event, user: str = Depends(authenticate)) -> dict:
+    body.creator = user
     await event_database.save(body)
     return {"message": "Event created successfully"}
 
 @event_router.put("/edit/{id}", response_model=Event)
-async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
+async def update_event(id: PydanticObjectId, body: EventUpdate, user: str = Depends(authenticate)) -> Event:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation not allowed"
+        )
     updated_event = await event_database.update(id, body)
     if not update_event:
         raise HTTPException(
@@ -41,13 +49,20 @@ async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
     return updated_event
 
 @event_router.delete("/delete/{id}")
-async def delete_event(id: PydanticObjectId) -> dict:
-    event = await event_database.delete(id)
+async def delete_event(id: PydanticObjectId, user: str = Depends(authenticate)) -> dict:
+    event = await event_database.get(id)
     if not event:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
-            detail="Event with supplied ID does not exist"
+            detail="Event not found"
         )
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Operation not allowed"
+        )
+    event = await event_database.delete(id)
+    
     return {"message": "Event deleted successfully"}
 
 @event_router.delete("/")
